requirements.txt
tensorflow-gpu==2.2.0
Pillow==7.1.2
numpy==1.20.3
opencv-python==4.2.0.32
opencv-contrib-python==4.2.0.32
